added not managed customers inside dashboard
