+++ components/queueManager/Index.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2023 Nethesis S.r.l.
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-export * from './QueueManagerDashboard'
+export * from './Dashboard/QueueManagerDashboard'
 export * from './QueueManagement'
 export * from './NotManagedCalls'
 export * from './NotManagedCallsFilter'
