+++ lib/queueManager.tsx
@@ -587,7 +587,6 @@ export const DEFAULT_REALTIME_FIRST_QUEUE_SELECTED = {}
 export const DEFAULT_REALTIME_SECOND_QUEUE_SELECTED = {}
 export const DEFAULT_REALTIME_IS_SECOND_CARD_VISIBLE = false
 
-
 // Set default queue manager tab to dashboard
 export const DEFAULT_SELECTED_TAB_QUEUE_MANAGER = 'dashboard'
