+++ pages/queueManager.tsx
@@ -5,7 +5,7 @@ import type { NextPage } from 'next'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import classNames from 'classnames'
-import { QueueManagerDashboard } from '../components/queueManager/QueueManagerDashboard'
+import { QueueManagerDashboard } from '../components/queueManager/Dashboard/QueueManagerDashboard'
 import { QueueManagement } from '../components/queueManager/QueueManagement'
 import { NotManagedCalls } from '../components/queueManager/NotManagedCalls'
 import { RealTimeManagement } from '../components/queueManager/RealTimeManagement'
